Fix Broadcom RPC build following syncd container upgrade to Bookworm (#20949)
Why I did it
With the syncd container upgrade to Bookworm in #19712, the RPC containers started failing to build. There are some additional changes needed to the RPC container build and to use a newer version of the PTF package.

Work item tracking
Microsoft ADO (number only): 30368269
How I did it
Modify the RPC dockerfiles to use Python 3, and to use the existing nanomsg package in Debian.

Also, enable the Broadcom RPC image build as part of PRs.