Build and include tacacs+ for PTF Py3 only image (#20701)
Why I did it
The "tacacs+" package is not available in the apt repository for Debian Bullseye and beyond. The sonic-mgmt tests require tacacs+ and this PR builds it from source and adds it to the PY3 only image.

Work item tracking
Microsoft ADO (number only): 29946202

How I did it
Clone tacacs+ from https://github.com/robot527/tac_plus build and install it.

How to verify it
By running sonic-mgmt restart-ptf step with newly built image.

Tested branch (Please provide the tested image version)
Not applicable.

Description for the changelog
[docker-ptf]: Build and include tacacs+ for py3only image

The "tacacs+" package is not available in the apt repository for Debian Bullseye and beyond. The sonic-mgmt tests require tacacs+ and this commit builds it from source and adds it to the PY3 only image.
Link to config_db schema for YANG module changes
Not applicable.